requirements.txt
# Azure AI and Authentication
azure-identity
azure-ai-agents==1.2.0b1
azure-ai-projects
azure-core-tracing-opentelemetry

# OpenTelemetry and Monitoring
azure-monitor-opentelemetry
opentelemetry-sdk
opentelemetry-exporter-otlp

# Web UI and Environment
streamlit>=1.40.0
python-dotenv

# Testing
pytest>=7.0.0

# Utilities and Type Support
watchdog>=6.0.0
dataclasses-json>=0.6.3
typing-extensions>=4.5.0
pydantic>=2.0.0
bleach>=6.0.0
